python-collector/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1
lxml==5.1.0
pytz==2024.1
openpyxl==3.1.2
